+++ src/index.ts
@@ -1,6 +1,7 @@
-export { transpile } from './transpiler/index.js'
+export { transpile, transpileText } from './transpiler/index.js'
 export {
   WorkflowSyntaxError,
+  WorkflowSyntaxErrorWithText,
   InternalTranspilingError,
   SourceCodeLocation,
 } from './errors.js'