+++ test/transpiler.spec.ts
@@ -2,7 +2,7 @@ import { expect } from 'chai'
 import * as YAML from 'yaml'
 import * as fs from 'node:fs'
 import { transpile, transpileText } from '../src/transpiler/index.js'
-import { assertTranspiled, assertNotCalled } from './testutils.js'
+import { assertTranspiled } from './testutils.js'
 import { IOError, WorkflowSyntaxError } from '../src/errors.js'
 
 describe('Type annotations', () => {
@@ -426,13 +426,10 @@ describe('Sample source files', () => {
     fs.readdirSync(samplesdir).forEach((file) => {
       if (file.endsWith('.ts')) {
         const fullPath = `${samplesdir}/${file}`
-        const input = {
-          filename: fullPath,
-          read: assertNotCalled,
-        }
+        const sourceCode = fs.readFileSync(fullPath, 'utf-8')
 
         expect(() =>
-          transpile(input, 'samples/tsconfig.json', false),
+          transpile(fullPath, sourceCode, 'samples/tsconfig.json', false),
         ).not.to.throw()
       }
     })
@@ -442,23 +439,18 @@ describe('Sample source files', () => {
     fs.readdirSync(samplesdir).forEach((file) => {
       if (file.endsWith('.ts')) {
         const fullPath = `${samplesdir}/${file}`
-        const input = {
-          filename: fullPath,
-          read: () => fs.readFileSync(fullPath, 'utf-8'),
-        }
+        const sourceCode = fs.readFileSync(fullPath, 'utf-8')
 
-        expect(() => transpile(input, undefined, false)).not.to.throw()
+        expect(() =>
+          transpile(fullPath, sourceCode, undefined, false),
+        ).not.to.throw()
       }
     })
   })
 
   it('generates linked output', () => {
     const fullPath = `${samplesdir}/sample2.ts`
-    const input = {
-      filename: fullPath,
-      read: assertNotCalled,
-    }
-    const yaml = transpile(input, 'samples/tsconfig.json', true)
+    const yaml = transpile(fullPath, '', 'samples/tsconfig.json', true)
     const observed = YAML.parse(yaml) as Record<string, unknown>
 
     // main comes from sample2.ts
@@ -467,12 +459,9 @@ describe('Sample source files', () => {
   })
 
   it('throws if the input file does not exist', () => {
-    const input = {
-      filename: `${samplesdir}/this-file-does-not-exist.ts`,
-      read: assertNotCalled,
-    }
+    const fullPath = `${samplesdir}/this-file-does-not-exist.ts`
     const transpileAttempt = () =>
-      transpile(input, 'samples/tsconfig.json', false)
+      transpile(fullPath, '', 'samples/tsconfig.json', false)
 
     expect(transpileAttempt).to.throw(IOError, 'not found')
   })